# Replace triple dot speed optimization for newer ruby versions: