@Edouard-chin 2025-03-14
Oops, I forgot 🙈 . It's the same as the one from the [ruby ticket](https://bugs.ruby-lang.org/issues/19165). I stripped other comparisons (splat, double splat..)

```ruby
require 'forwardable'
require 'pp'
require 'benchmark/ips'

class Obj
  extend Forwardable

  attr_accessor :other

  def initialize
    @other = Other.new
  end

  def foo_with_triple_dots(...)
    @other.foo(...)
  end

  delegate :foo => :@other
end

class Other
  def foo() end
end

o = Obj.new

Benchmark.ips do |x|
  x.report 'simple call' do
    o.other.foo
  end

  x.report 'triple dots' do
    o.foo_with_triple_dots
  end

  x.compare!
end

```