@byroot 2025-03-14
I think we should craft a more dedicated benchmark, and also try it with YJIT.

Also, given the work was done already, even if 1.5x isn't that much, we might as well keep using the faster code, but if it's very close on 3.5, we can start getting rid of it for that version.