@Edouard-chin 2025-03-14
Make sense. I'll check with a more specific benchmark and try with yjit. Will post the result and close this PR it it end up being similar on 3.5