@byroot 2025-03-14
Thanks.

Also I think

> 3\. Tools that introspect method signature

Has some value. I don't use any of these tools, so not sure how valuable that really is? The only downside being the explicit block delegation, but other than that I like the idea of the method signature being representative.