@Edouard-chin 2025-03-14
I run a few more benchmarks with a more dedicated script on multiple ruby with and without YJIT. In summary, starting from Ruby 3.4 the speed improvement for `...` **without yjit** almost matches the performance of passing parameters manually. 
**Without yjit** is still 3x slower. 

I'll close this PR for this reason and I also agree with the "but other than that I like the idea of the method signature being representative."

Here are the benchmarks:

---------------------------

<details>
<summary> ✏️ Script </summary>

```ruby
require "bundler/inline"

gemfile(true) do
  source "https://rubygems.org"

  gem "activesupport", '8.0.1'
  gem "benchmark-ips"
end

require "active_support"
require "active_support/core_ext/module/delegation.rb"
require 'benchmark/ips'

class Other
  def handcrafted_signature(a) end
  def triple_dots_forwarding(key:) end
end

class Obj
  attr_accessor :other

  def initialize
    @other = Other.new
  end

  ActiveSupport::Delegation.generate(self, [:handcrafted_signature, :triple_dots_forwarding], to: :@other, as: Other)
end

o = Obj.new
puts "Signature for 'handcrafted_signature' method: #{o.method(:handcrafted_signature).parameters}"
puts "Signature for 'triple_dots_forwarding' method:  #{o.method(:triple_dots_forwarding).parameters}\n\n"

Benchmark.ips do |x|
  x.report 'delegated with handcrafted signature' do
    o.handcrafted_signature(1)
  end

  x.report 'delegated with triple dots' do
    o.triple_dots_forwarding(key: 1)
  end

  x.compare!
end

```
</details>

<details>
<summary> ♦️ Ruby 3.3.6. No YJIT. (~4x slower) </summary>

```
Signature for 'handcrafted_signature' method: [[:req, :a], [:block, :&]]
Signature for 'triple_dots_forwarding' method:  [[:rest, :*], [:keyrest, :**], [:block, :&]]

ruby 3.3.6 (2024-11-05 revision 75015d4c1f) [x86_64-darwin22]
Warming up --------------------------------------
delegated with handcrafted signature
                       819.456k i/100ms
delegate with triple dots
                       186.410k i/100ms
Calculating -------------------------------------
delegated with handcrafted signature
                          7.396M (±10.1%) i/s  (135.21 ns/i) -     36.876M in   5.039480s
delegate with triple dots
                          1.771M (± 7.1%) i/s  (564.78 ns/i) -      8.948M in   5.080278s

Comparison:
delegated with handcrafted signature:  7396034.9 i/s
delegate with triple dots:  1770610.0 i/s - 4.18x  slower
```
</details>

<details>
<summary> ♦️ Ruby 3.4.1 No YJIT. (1.16x slower) </summary>

```
Signature for 'handcrafted_signature' method: [[:req, :a], [:block, :&]]
Signature for 'triple_dots_forwarding' method:  [[:rest, :*], [:keyrest, :**], [:block, :&]]

ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +PRISM [x86_64-darwin22]
Warming up --------------------------------------
delegated with handcrafted signature
                       708.298k i/100ms
delegated with triple dots
                       648.135k i/100ms
Calculating -------------------------------------
delegated with handcrafted signature
                          7.787M (± 4.0%) i/s  (128.41 ns/i) -     38.956M in   5.010697s
delegated with triple dots
                          6.703M (± 4.1%) i/s  (149.18 ns/i) -     33.703M in   5.036267s

Comparison:
delegated with handcrafted signature:  7787356.5 i/s
delegated with triple dots:  6703415.4 i/s - 1.16x  slower
```
</details>

<details>
<summary> ♦️ Ruby 3.4.1. With YJIT (~3x slower) </summary>

```
Signature for 'handcrafted_signature' method: [[:req, :a], [:block, :&]]
Signature for 'triple_dots_forwarding' method:  [[:rest, :*], [:keyrest, :**], [:block, :&]]

ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +YJIT +PRISM [x86_64-darwin22]
Warming up --------------------------------------
delegated with handcrafted signature
                         1.816M i/100ms
delegated with triple dots
                       584.927k i/100ms
Calculating -------------------------------------
delegated with handcrafted signature
                         20.684M (± 3.4%) i/s   (48.35 ns/i) -    103.505M in   5.010546s
delegated with triple dots
                          6.544M (± 4.9%) i/s  (152.81 ns/i) -     32.756M in   5.018088s

Comparison:
delegated with handcrafted signature: 20683877.0 i/s
delegated with triple dots:  6544097.4 i/s - 3.16x  slower
```
</details>

➡️ Results on Ruby master, with and without YJIT are similar to 3.4.1